requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
ecdsa==0.14.1
fastapi==0.63.0
gitdb==4.0.5
GitPython==3.1.12
h11==0.12.0
idna==3.1
multidict==5.1.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
rsa==4.7
six==1.15.0
smmap==3.0.5
SQLAlchemy==1.3.23
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.3
yarl==1.6.3
